+++ controller/joblistcontroller.php
@@ -111,6 +111,10 @@ class JobListController extends Controller {
 	 * @return void
 	 */
 	private function checkEditorsCheckJob() {
+		if (!$this->config->getCronChecker()) {
+			$this->removeJob(EditorsCheck::class);
+			return;
+		}
 		if ($this->config->getEditorsCheckInterval() > 0) {
 			$this->addJob(EditorsCheck::class);
 		} else {